examples/rag/qdrant/requirements.txt
fastembed>=0.1.0
qdrant-client>=1.7.0
langchain>=0.1.0
langchain-core>=0.1.10
langchain-openai>=0.0.5
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.6.0
unstructured>=0.10.30
pdf2image>=1.16.3
pypdf>=3.17.1 